Segunda versión del portafolio - se agrega el navBar - se agrega el header - se agrega un elemnto item para tener una guia:::
