+++ components/header/header.css
@@ -1,0 +1,20 @@
+.header {
+    width: 100%;
+
+    display: flex;
+    justify-content: space-between;
+}
+
+.header h2 {
+    font-size: 2rem;
+}
+
+
+.header div {
+    background-color: black;
+    color: white;
+    padding: 0.5rem 1rem;
+    border-radius: 3px;
+    font-size: 1.4rem;
+    font-weight: 900x;
+}
