+++ components/header/header.js
@@ -1,0 +1,18 @@
+function header(){
+    let div = document.createElement('div');
+    div.className = 'header';
+    
+    let h2 = document.createElement('h2');
+    h2.innerText = "Proyectos";
+    div.appendChild(h2);
+
+    let btn = document.createElement('div');
+    btn.innerHTML = "Github";
+    div.appendChild(btn);
+
+
+    return div;
+}
+
+
+export { header }
