+++ components/project/project.css
@@ -2,23 +2,3 @@
     padding: 1.5rem 2rem;
 }
 
-.header {
-    width: 100%;
-
-    display: flex;
-    justify-content: space-between;
-}
-
-.header h2 {
-    font-size: 2rem;
-}
-
-
-.header div {
-    background-color: black;
-    color: white;
-    padding: 0.5rem 1rem;
-    border-radius: 3px;
-    font-size: 1.4rem;
-    font-weight: 900x;
-}
