+++ components/project/project.js
@@ -1,3 +1,5 @@
+import { header } from "../header/header.js"
+
 function proyectos(){
 
     let section = document.createElement('section');
@@ -9,21 +11,7 @@ function proyectos(){
     return section;
 }
 
-function header(){
-    let div = document.createElement('div');
-    div.className = 'header';
-    
-    let h2 = document.createElement('h2');
-    h2.innerText = "Proyectos";
-    div.appendChild(h2);
 
-    let btn = document.createElement('div');
-    btn.innerHTML = "Github";
-    div.appendChild(btn);
-
-
-    return div;
-}
 
 function item(){
     let div = document.createElement('div');
