+++ index.html
@@ -7,6 +7,7 @@
 
     <!-- css -->
      <link rel="stylesheet" href="components/navBar/navBar.css">
+     <link rel="stylesheet" href="components/header/header.css">
      <link rel="stylesheet" href="components/project/project.css">
      <link rel="stylesheet" href="index.css">
      <link rel="stylesheet" href="config/config.css">
